+++ src/main/groovy/com/cedarsoftware/controller/NCubeController.groovy
@@ -140,21 +140,31 @@ class NCubeController implements NCubeConstants, RpmVisualizerConstants
         return mutableClient.userId
     }
 
-    Boolean assertPermissions(ApplicationID appId, String resource, Action action)
+    Boolean assertPermissions(ApplicationID appId, String resource, String actionName)
     {
         appId = addTenant(appId)
+        Action action = actionName == null ? null : Action.valueOf(actionName.toUpperCase())
         return mutableClient.assertPermissions(appId, resource, action)
     }
 
     Map checkMultiplePermissions(ApplicationID appId, String resource, Object[] actions)
     {
+        if (ArrayUtilities.isEmpty(actions))
+        {
+            throw new IllegalArgumentException('Must send at least action to check permissions')
+        }
         appId = addTenant(appId)
+        for (int i = 0; i < actions.length; i++)
+        {
+            actions[i] = Action.valueOf((actions[i] as String).toUpperCase())
+        }
         return mutableClient.checkMultiplePermissions(appId, resource, actions)
     }
 
-    Boolean checkPermissions(ApplicationID appId, String resource, String action)
+    Boolean checkPermissions(ApplicationID appId, String resource, String actionName)
     {
         appId = addTenant(appId)
+        Action action = actionName == null ? null : Action.valueOf(actionName.toUpperCase())
         return mutableClient.checkPermissions(appId, resource, action)
     }
 
@@ -395,7 +405,13 @@ class NCubeController implements NCubeConstants, RpmVisualizerConstants
         verifyAllowExecute('mapReduce')
         appId = addTenant(appId)
         NCube ncube = getCubeInternal(appId, cubeName)
-        return ncube.mapReduce(rowAxisName, colAxisName, where, input, output, columnsToSearch, columnsToReturn)
+        GroovyShell shell = new GroovyShell()
+        Object whereClosure = shell.evaluate(where)
+        if (!(whereClosure instanceof Closure))
+        {
+            throw new IllegalArgumentException("Passed in 'where' clause: ${where}, is not evaluating to a Closure.  Make sure it is in the form (example): { Map input -> input.state == 'AZ' }")
+        }
+        return ncube.mapReduce(rowAxisName, colAxisName, (Closure)whereClosure, input, output, columnsToSearch, columnsToReturn)
     }
 
     private static Map<String, CellInfo> valuesToCellInfo(Map<String, Object> metaProps)
@@ -1004,11 +1020,11 @@ class NCubeController implements NCubeConstants, RpmVisualizerConstants
         mutableClient.updateCube(ncube)
     }
     
-    void createReferenceFromAxis(ApplicationID appId, String cubeName, String axisName, ApplicationID refAppId, String refCubeName, String refAxisName)
+    void createRefAxis(ApplicationID appId, String cubeName, String axisName, ApplicationID refAppId, String refCubeName, String refAxisName)
     {
         appId = addTenant(appId)
         refAppId = addTenant(refAppId)
-        mutableClient.createReferenceFromAxis(appId, cubeName, axisName, refAppId, refCubeName, refAxisName)
+        mutableClient.createRefAxis(appId, cubeName, axisName, refAppId, refCubeName, refAxisName)
     }
 
     Boolean renameCube(ApplicationID appId, String oldName, String newName)
@@ -1789,8 +1805,17 @@ class NCubeController implements NCubeConstants, RpmVisualizerConstants
         // App server name and version
         Map serverStats = [:]
 
+        String username
+        try
+        {
+            username = mutableClient.userId
+        }
+        catch (Exception ignored)
+        {
+            username = '--'
+        }
         putIfNotNull(serverStats, 'n-cube version', versionNumber)
-        putIfNotNull(serverStats, 'User ID', mutableClient.userId)
+        putIfNotNull(serverStats, 'User ID', username)
         putIfNotNull(serverStats, 'Java version', getAttribute(mbs, 'JMImplementation:type=MBeanServerDelegate', 'ImplementationVersion'))
         putIfNotNull(serverStats, 'hostname, servlet', getServletHostname())
         putIfNotNull(serverStats, 'hostname, OS', getInetHostname())
@@ -1845,7 +1870,7 @@ class NCubeController implements NCubeConstants, RpmVisualizerConstants
         {
             Map.Entry<String, Object> entry = it.next()
             String key = entry.key
-            if (!(key.startsWith("heap") || key.startsWith('nonheap') || key.startsWith('processors') || key.startsWith('mem')))
+            if (!(key.startsWith('heap') || key.startsWith('nonheap') || key.startsWith('processors') || key.startsWith('mem')))
             {
                 if (showAll)
                 {