+++ src/test/groovy/com/cedarsoftware/ncube/TestNCubeManager.groovy
@@ -1526,11 +1526,19 @@ class TestNCubeManager extends NCubeCleanupBaseTest
     void testCopyBranch()
     {
         ApplicationID copyAppId = defaultSnapshotApp.asBranch('copy')
+        ApplicationID copyAppId2 = defaultSnapshotApp.asBranch('copy2')
         NCube cube = createCubeFromResource(defaultSnapshotApp, 'latlon.json')
 
         mutableClient.copyBranch(defaultSnapshotApp, copyAppId)
-        NCube copiedCube = mutableClient.getCube(copyAppId, cube.name)
-        assertNotNull(copiedCube)
+        List<NCubeInfoDto> dtos = mutableClient.search(copyAppId, cube.name, null, [(SEARCH_INCLUDE_NOTES):true])
+        assert 1 == dtos.size()
+        assertContainsIgnoreCase(dtos[0].notes, copyAppId.toString(), 'copied from', defaultSnapshotApp.toString())
+
+        mutableClient.copyBranch(copyAppId, copyAppId2)
+        dtos = mutableClient.search(copyAppId2, cube.name, null, [(SEARCH_INCLUDE_NOTES):true])
+        assert 1 == dtos.size()
+        assertContainsIgnoreCase(dtos[0].notes, copyAppId2.toString(), 'copied from', copyAppId.toString())
+        assert -1 == dtos[0].notes.indexOf(defaultSnapshotApp.toString())
     }
 
     @Test
@@ -1828,12 +1836,12 @@ class TestNCubeManager extends NCubeCleanupBaseTest
         assert prId
 
         // assert admin can read prcube
-        assert mutableClient.checkPermissions(sysAppId, "tx.${prId}".toString(), Action.READ.name())
+        assert mutableClient.checkPermissions(sysAppId, "tx.${prId}".toString(), Action.READ)
 
         // assert other user can't read prcube
         NCubeManager manager = NCubeAppContext.getBean(MANAGER_BEAN) as NCubeManager
         manager.userId = 'otherUser'
-        assert !mutableClient.checkPermissions(sysAppId, "tx.${prId}".toString(), Action.READ.name())
+        assert !mutableClient.checkPermissions(sysAppId, "tx.${prId}".toString(), Action.READ)
 
         try
         {   // make sure other user can still merge the pr (ignoring permissions)